# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |-
    The user is presented with a list of conversations with other users or with groups, sorted in reverse
    chronological order. Each element in the list must display the username of the other person or the
    group name, the user profile photo or the group photo, the date and time of the latest message,
    the preview (snippet) of the text message, or an icon for a photo message. The user can start a new
    conversation with any other user of WASAText, and this conversation will automatically be added
    to the list. The user can search for other users via the username and see all the existing WASAText
    usernames.
    The user can create a new group with any number of other WASAText users to start a conversation.
    Group members can add other users to the group, but users cannot join groups on their own or even
    see groups they aren’t a part of. Additionally, users have the option to leave a group at any time.
    The user can open a conversation to view all exchanged messages, displayed in reverse chronological
    order. Each message includes the timestamp, the content (whether text or photo), and the sender’s
    username for received messages, or one/two checkmarks to indicate the status of sent messages. Any
    reactions (comments) on messages are also displayed, along with the names of the users who posted
    them.
...
    recipient (by all the recipients for groups) within the conversation itself.
...
servers:
...
paths:
...
  /users/{user_id}/name:
    parameters:
      - $ref: "#/components/parameters/UserId"
    put:
      security:
        - bearerAuth: []
      tags: ["user"]
      summary: Updates the user name
      description: |-
        Updates the user name with the new one.
      operationId: setMyUserName
      requestBody:
        description: New name for the user.
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Object containing the new user name.
              properties:
                name:
                  $ref: "#/components/schemas/Username"
              required:
                - name
      responses:
        '200':
          description: User name updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Confirmation message.
                properties:
                  message:
                    type: string
                    description: Confirmation message.
                    pattern: "^.*$"
                    minLength: 1
                    maxLength: 999
              example:
                message: "User name updated"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /users/{user_id}/photo:
    parameters:
      - $ref: "#/components/parameters/UserId"
    put:
      security:
        - bearerAuth: []
      tags: ["user"]
      summary: Updates the user photo
      description: |-
        Updates the user photo with the new one.
      operationId: setMyPhoto
      requestBody:
        description: New photo for the user.
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Object containing the new user photo.
              properties:
                photo:
                  $ref: "#/components/schemas/Photo"
              required:
                - photo
      responses:
        '200':
          description: User photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Confirmation message.
                properties:
                  message:
                    type: string
                    description: Confirmation message.
                    pattern: "^.*$"
                    minLength: 1
                    maxLength: 999
              example:
                message: "User photo updated"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /users:
    parameters:
      - $ref: "#/components/parameters/SearchUsers"
    get:
      tags: ["user"]
      summary: Search for users by username
      description: |-
        Returns a list of users that match with the username.
      operationId: searchUsers
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: Users that match with the username.
                properties:
                  userlist:
                    description: List of users.
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                    minItems: 0
                    maxItems: 999
                required:
                  - userlist
              example:
                userlist: []
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
...
    UserId:
      name: user_id
      in: path
      description: ID of the user.
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 2147483647
        example: 1
        readOnly: true
...
    SearchUsers:
      name: search_users
      in: query
      description: Username to search for.
      required: true
      schema:
        type: string
        example: "user1"
        pattern: "^.*$"
        minLength: 0
        maxLength: 16
  schemas:
    Username:
      title: username
      description: Username of the user.
      type: string
      example: "user1"
      pattern: "^.*$"
      minLength: 3
      maxLength: 16
...
    Photo:
      title: photo
      description: Photo of the user or group.
      type: string
      format: url
      example: "https://example.com/photo.jpg"
      pattern: "^.*$"
      minLength: 1
      maxLength: 999
    UserId:
      title: userid
      description: ID of the user.
      type: integer
      minimum: 1
      maximum: 2147483647
      example: 1
...
    User:
      title: user
      description: User details.
      type: object
      properties:
        userid:
          $ref: "#/components/schemas/UserId"
        username:
          $ref: "#/components/schemas/Username"
      required:
        - userid
        - username
...
  responses:
    BadRequest:
      description: The request was not compliant with the documentation.
      content:
        application/json:
          schema:
            type: object
            description: Error response object.
            properties:
              error:
                type: string
                description: Error message.
                pattern: "^.*$"
                minLength: 1
                maxLength: 200
          example:
            error: "Bad Request"
    Unauthorized:
      description: Access token is missing or invalid.
      content:
        application/json:
          schema:
            type: object
            description: Error response object.
            properties:
              error:
                type: string
                description: Error message.
                pattern: "^.*$"
                minLength: 1
                maxLength: 200
          example:
            error: "Unauthorized"
    NotFound:
      description: The server cannot find the requested resource.
      content:
        application/json:
          schema:
            type: object
            description: Error response object.
            properties:
              error:
                type: string
                description: Error message.
                pattern: "^.*$"
                minLength: 1
                maxLength: 200
          example:
            error: "Not Found"
    InternalServerError:
      description: The server encountered an unexpected condition that prevented it from fulfilling the request.
      content:
        application/json:
          schema:
            type: object
            description: Error response object.
            properties:
              error:
                type: string
                description: Error message.
                pattern: "^.*$"
                minLength: 1
                maxLength: 200
          example:
            error: "Internal Server Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer